Unified Bonus withdrawals Process Page.
The Idea is to create a unique withdrawal transaction from our admin for all admin staff to be able to generate the amount they get as unifed bonus on our system.

And This should not reflect in our saldo. but the record will be treated and kept within the system as a normal withdrawal transaction.